refactor: extract IssueExportCompletion.swift from IssueExportSupport.swift
Splits the `IssueExportCompletion` result value out.

Byte-identical move of lines 15-29. Both files import Foundation.

IssueExportSupport.swift: 29 → 14 lines. Tests: 667.
