+++ Sources/BugNarrator/Services/IssueExportSupport.swift
@@ -12,18 +12,3 @@ struct IssueExportRequestContext {
     let apiKey: String
 }
 
-struct IssueExportCompletion {
-    let destination: ExportDestination
-    let sessionID: UUID
-    let results: [ExportResult]
-    let duplicateCount: Int
-    let performedRemoteExport: Bool
-
-    var summary: String {
-        IssueExportReviewPolicy.exportSummary(
-            for: results,
-            duplicateCount: duplicateCount,
-            destination: destination
-        )
-    }
-}
